add confirm del

+++ src/views/ContactManager.vue
@@ -63,7 +63,7 @@
                 <!-- v-for="contact in sortedContacts" -->
                 <!-- <div class="col-md-4" v-for="contact in sortedContacts" :key="contact.id"> -->
                 <div class="col-md-4" v-for="contact in sortedContacts" :key="contact.fields.id">
-                <!-- <div class="col-md-4" v-for="contact in contacts" :key="contact.fields.id"> -->
+                    <!-- <div class="col-md-4" v-for="contact in contacts" :key="contact.fields.id"> -->
                     <div class="card my-2 list-group-item-success shadow-lg">
                         <div class="card-body">
                             <div class="row align-items-center">
@@ -92,10 +92,12 @@
                                 </div>
                                 <div class="col-sm-1">
                                     <!-- <div class="col-sm-1 d-flex flex-column justify-content-center align-items-center"> -->
-                                    <router-link :to="`/contacts/view/${contact.fields.id}`" class="btn btn-warning my-1 me-2">
+                                    <router-link :to="`/contacts/view/${contact.fields.id}`"
+                                        class="btn btn-warning my-1 me-2">
                                         <i class="fa fa-eye"></i>
                                     </router-link>
-                                    <router-link :to="`/contacts/edit/${contact.fields.id}`" class="btn btn-primary my-1 me-2">
+                                    <router-link :to="`/contacts/edit/${contact.fields.id}`"
+                                        class="btn btn-primary my-1 me-2">
                                         <i class="fa fa-pen"></i>
                                     </router-link>
                                     <button class="btn btn-danger my-1 me-2" @click="myDelete(contact.fields.id)">
@@ -214,20 +216,22 @@ export default {
             this.currentUserId = await this.currentUser();
 
             if (this.currentUserId) {
-                try {
-                    this.loading = true
-                    let response = await ContactService.deleteCondo(itemId, this.currentUserId)
-                    if (response) {
-                        console.log('deleted')
-                        // const airtableId = recordId.toString();
-                        let response = await ContactService.getAllCondos(this.currentUserId)
-                        this.contacts = response.data.records
+                if (confirm("Confirm delete?")) {
+                    try {
+                        this.loading = true
+                        let response = await ContactService.deleteCondo(itemId, this.currentUserId)
+                        if (response) {
+                            console.log('deleted')
+                            // const airtableId = recordId.toString();
+                            let response = await ContactService.getAllCondos(this.currentUserId)
+                            this.contacts = response.data.records
+                            this.loading = false
+                        }
+                    }
+                    catch (error) {
+                        this.errorMessage = error
                         this.loading = false
                     }
-                }
-                catch (error) {
-                    this.errorMessage = error
-                    this.loading = false
                 }
             }
         },
@@ -278,5 +282,4 @@ export default {
     white-space: nowrap;
     max-width: 200px;
     /* adjust this value to change the maximum width */
-}
-</style>
+}</style>
